api/requirements.txt
numpy
opencv-contrib-python
imutils
pydrive
python-dotenv
gevent
fastapi
uvicorn